work around a ghc-8.4.1 bug (see #58)
When a module is loaded by file name, ghc-8.4.1 loses track of the file
location after the first time it has been loaded, which breaks hint's
treatment of phantom modules.

With this patch, hint creates a new temporary subdirectory for storing
the phantom modules, and adds that to ghc's search path. The phantom
module will be created as M<nnn>.hs as a file name instead of just <nnn>,
so that ghc actually finds the module.

The latter change also affects previous ghc versions; this means that if
hint leaves behind temporary files, then those will now be called
M<nnn>.hs rather than just <nnn>. This seems to be a good idea because
this way they will be easier to identify as Haskell files.