ensure that source file is removed when removing a phantom module
When removing a phantom module, hint checks whether it is safe to remove;
if so it removes it from the list of targets and reloads all modules,
before removing the phantom module's source file.

If the reload step fails (for example because another module fails to
load), then the source file was never deleted; this patch ensure that
the file is always removed.

With this change, the testsuite passes for ghc 8.4.1.