CI: start caching ~/.cabal/, ~/.ghc/, and dist/
This should hopefully lead to much faster build times. In practive
however, it only seems to save us about a minute from the original 12 or
so.

While at it, simplify the series of cabal commands to run. Even with
caching, cabal can take a non-trivial time to realise it doesn't need to
do anything.