ifInterpreterNotRunning should have a consistent meaning
Namely, it should always throw an exception if another instance of the
interpreter is already running. The reason this function exists is that
ghc's linker used not to be thread-safe, and so it was better to throw
an exception early on than to let the user encounter weird errors later
on.

When THREAD_SAFE_LINKER is defined, there is no reason to throw such an
exception, but the proper way to not throw that exception is to not call
ifInterpreterNotRunning, not to change the meaning of ifInterpreterNotRunning.